`@0x:contracts-exchange` Added tests for fillOrder protocol fees
