+++ contracts/exchange/8:p
@@ -1,0 +1,29 @@
+import { blockchainTests, constants } from '@0x/contracts-test-utils';
+
+import { artifacts, ExchangeContract, TestProtocolFeesContract, TestProtocolFeesReceiverContract } from '../src';
+
+blockchainTests('Protocol Fee Payments', env => {
+    let testProtocolFees: TestProtocolFeesContract;
+    let testProtocolFeesReceiver: TestProtocolFeesReceiverContract;
+
+    before(async () => {
+        testProtocolFees = await TestProtocolFeesContract.deployFrom0xArtifactAsync(
+            artifacts.TestProtocolFees,
+            env.provider,
+            env.txDefaults,
+            {},
+        );
+        testProtocolFeesReceiver = await TestProtocolFeesReceiverContract.deployFrom0xArtifactAsync(
+            artifacts.TestProtocolFeesReceiver,
+            env.provider,
+            env.txDefaults,
+            {},
+        );
+    });
+
+    blockchainTests.resets('fillOrder Protocol Fees', () => {
+        it('should pay protocol fee in ETH when the correct value is sent', async () => {
+            await testProtocolFeesReceiver.testFillOrderProtocolFees.awaitTransactionSuccessAsync(;
+        });
+    });
+});
