+++ contracts/exchange/tsconfig.json
@@ -32,6 +32,8 @@
         "generated-artifacts/TestAssetProxyDispatcher.json",
         "generated-artifacts/TestExchangeInternals.json",
         "generated-artifacts/TestLibExchangeRichErrorDecoder.json",
+        "generated-artifacts/TestProtocolFees.json",
+        "generated-artifacts/TestProtocolFeesReceiver.json",
         "generated-artifacts/TestSignatureValidator.json",
         "generated-artifacts/TestTransactions.json",
         "generated-artifacts/TestValidatorWallet.json",
